update: hooks/apiの中の構成変更
