+++ app/line-connect/liff/messages/page.tsx
@@ -18,7 +18,7 @@ import { getWorkSessionsByUserId } from "@/lib/api/workSession";
 import { messageApi, CreateMessageParams } from "@/lib/api/message";
 import useSWR from "swr";
 import { LinkAccountScreen } from "../components/LinkAccountScreen";
-import { useSubscriptionMessagesByUserId } from "@/hooks/api/messages/useSubscriptionMessagesByUserId";
+import { useSubscriptionMessagesByUserId } from "@/hooks/api/user/messages/useSubscriptionMessagesByUserId";
 
 interface AnimatedSectionProps {
   children: React.ReactNode;
