+++ hooks/api/companyuser/companyUsers/useGetCompanyUsersByCompanyId.ts
@@ -1,16 +1,16 @@
 import { getApi } from "@/api/api-factory"
 import useSWR from "swr"
-import { QueryConfigType } from "../config-type";
+import { QueryConfigType } from "../../config-type";
 import { Companies } from "@/api/__generated__/base/Companies";
 
 export interface Params {
   companyId?: string;
 }
 
-export const useGetRestaurantsByCompanyId = (params: Params, config?: QueryConfigType) => {
+export const useGetCompanyUsersByCompanyId = (params: Params, config?: QueryConfigType) => {
   const { dedupingInterval } = config || {};
   const companies = getApi(Companies);
-  return useSWR(...companies.restaurantsListQueryArgs(params.companyId ?? '', {
+  return useSWR(...companies.companyusersListQueryArgs(params.companyId ?? '', {
     headers: {
       "X-User-Type": "company"
     }
