--- hooks/api/jobs/useGetJob.ts
@@ -1,20 +1,0 @@
-import { getApi } from "@/api/api-factory"
-import useSWR from "swr"
-import { QueryConfigType } from "../config-type";
-import { Jobs } from "@/api/__generated__/base/Jobs";
-
-export interface Params {
-  jobId?: number;
-}
-
-export const useGetJob = (params: Params, config?: QueryConfigType) => {
-  const { dedupingInterval } = config || {};
-  const jobs = getApi(Jobs);
-  return useSWR(...jobs.jobsDetailQueryArgs(params.jobId ?? -1, {
-    headers: {
-      "X-User-Type": "company"
-    }
-  }, params.jobId != null), {
-    dedupingInterval
-  });
-}
