+++ hooks/api/user/messages/useSubscriptionUnreadMessagesByUser.ts
@@ -1,8 +1,8 @@
 import { getApi } from "@/api/api-factory";
 import useSWR from "swr";
-import { QueryConfigType } from "../config-type";
+import { QueryConfigType } from "../../config-type";
 import { Chat } from "@/api/__generated__/base/Chat";
-import { UnreadMessage } from "./useSubscriptionUnreadMessagesByRestaurantId";
+import { UnreadMessage } from "../../companyuser/messages/useSubscriptionUnreadMessagesByRestaurantId";
 import useSWRSubscription from "swr/subscription";
 import realTimeClient from "@/api/xano";
 import { XanoRealtimeChannel } from "@xano/js-sdk/lib/models/realtime-channel";
