+++ hooks/api/user/worksessions/useGetWorksessionsByUserIdByTodo.ts
@@ -1,6 +1,6 @@
 import { getApi } from "@/api/api-factory"
 import useSWR from "swr"
-import { QueryConfigType } from "../config-type";
+import { QueryConfigType } from "../../config-type";
 import { Users } from "@/api/__generated__/base/Users";
 
 export interface Params {
